WiFiPeerListAdapterをOpponentListFragmentの内包クラスから独立したクラスに変更
