+++ app/src/main/java/jp/ac/it_college/std/shiritori/OpponentListFragment.java
@@ -4,7 +4,6 @@ package jp.ac.it_college.std.shiritori;
 import android.app.AlertDialog;
 import android.app.ListFragment;
 import android.app.ProgressDialog;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.NetworkInfo;
@@ -17,9 +16,7 @@ import android.provider.Settings;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.ListView;
-import android.widget.TextView;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -212,38 +209,4 @@ public class OpponentListFragment extends ListFragment
 
     }
 
-    private class WiFiPeerListAdapter extends ArrayAdapter {
-
-        private List<WifiP2pDevice> items;
-        private int resource;
-
-        public WiFiPeerListAdapter(Context context, int resource, List objects) {
-            super(context, resource, objects);
-            items = objects;
-            this.resource = resource;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            View v = convertView;
-            if (v == null) {
-                LayoutInflater vi = (LayoutInflater) getActivity().getSystemService(
-                        Context.LAYOUT_INFLATER_SERVICE);
-                v = vi.inflate(resource, null);
-            }
-            WifiP2pDevice device = items.get(position);
-            if (device != null) {
-                TextView top = (TextView) v.findViewById(R.id.device_name);
-                TextView bottom = (TextView) v.findViewById(R.id.device_details);
-                if (top != null) {
-                    top.setText(device.deviceName);
-                }
-                if (bottom != null) {
-                    bottom.setText(getDeviceStatus(device.status));
-                }
-            }
-
-            return v;
-        }
-    }
 }
