+++ app/src/main/java/jp/ac/it_college/std/shiritori/WiFiPeerListAdapter.java
@@ -1,0 +1,48 @@
+package jp.ac.it_college.std.shiritori;
+
+import android.content.Context;
+import android.net.wifi.p2p.WifiP2pDevice;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.TextView;
+
+import java.util.List;
+
+public class WiFiPeerListAdapter extends ArrayAdapter {
+
+    private List<WifiP2pDevice> items;
+    private int resource;
+
+    public WiFiPeerListAdapter(Context context, int resource, List objects) {
+        super(context, resource, objects);
+        items = objects;
+        this.resource = resource;
+    }
+
+    @Override
+    public View getView(int position, View convertView, ViewGroup parent) {
+        View v = convertView;
+        if (v == null) {
+            LayoutInflater vi = (LayoutInflater) getContext().getSystemService(
+                    Context.LAYOUT_INFLATER_SERVICE);
+            v = vi.inflate(resource, null);
+        }
+        WifiP2pDevice device = items.get(position);
+        if (device != null) {
+            TextView top = (TextView) v.findViewById(R.id.device_name);
+            TextView bottom = (TextView) v.findViewById(R.id.device_details);
+            if (top != null) {
+                top.setText(device.deviceName);
+            }
+            if (bottom != null) {
+//                bottom.setText(getDeviceStatus(device.status));
+            }
+        }
+
+        return v;
+    }
+
+
+}
